requirements.txt
# requirements.txt
streamlit==1.51.0
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11

requests
sqlparse
python-dotenv
pandas

matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
pillow>=10.0.0

# Optional API clients (uncomment if used)
# openai>=1.0.0
# anthropic
# google-generativeai
